Fix bug with Sorting/Filtering #33 - Bug appear when you try press filter, afrer sorting. Some films  dissappears. - solved by adding method getSortedAndFilteredFilms() in model, that can apply all modifications of films list together
